Add BusyBox ls (-l/-a/-R/-d); fix cross-boundary struct stat layout bug
ls.c is hand-written (not vendored) -- real upstream ls.c is ~1360 lines
of terminal-width-aware column output, 6 sort variants, and coloring that
doesn't fit this port's model and doesn't decompose into a small vendorable
core the way cp/mv/rm did. This covers the common case: -a (dotfiles),
-l (long format), -R (recurse), -d (list the dir itself), always one name
per line, sorted. Replaces the old app/ls.c demo app (same name, so it had
to go -- HBOS's hax dispatch is a flat name-keyed registry with no room
for two apps sharing a name).

While verifying -l, found every file reporting size 0 despite `cat` and
the shell's own (in-kernel, same-process) `stat` command showing the
correct size. Root cause: src/user/libc/syscall.h's struct stat used
`nlink_t = unsigned long` (8 bytes) while the kernel's struct stat
(src/sys/types.h) uses `nlink_t = uint32_t` (4 bytes) -- a mismatch that
shifts every field after st_nlink (uid/gid/rdev/size/atime/mtime/ctime)
by 4 bytes between what the kernel's stat()/fstat() syscall handlers
write and what user-mode code reads back. st_size ended up reading the
kernel's always-zero st_atime instead. This affected every user-mode
program calling stat()/fstat() and reading any field past st_nlink, not
just ls -- fixed by matching nlink_t's width to the kernel's.

Verified in QEMU: ls/-l/-a/-R/-d all behave correctly on nested dirs;
after the nlink_t fix, `ls -l`, `wc -c`, and `cp` all agree on file sizes
that previously only the shell's built-in `stat` reported correctly.

+++ src/user/libc/syscall.h
@@ -14,7 +14,16 @@ typedef unsigned int useconds_t;
 typedef long clock_t;
 typedef unsigned long dev_t;
 typedef unsigned long ino_t;
-typedef unsigned long nlink_t;
+/* Must be 4 bytes to match the kernel's struct stat (src/sys/types.h) --
+ * this used to be "unsigned long" (8 bytes), which shifted every field
+ * after st_nlink (uid/gid/rdev/size/atime/mtime/ctime) by 4 bytes between
+ * what the kernel's stat()/fstat() write and what user-mode code reads,
+ * so e.g. st_size always read back as 0 (actually reading half of the
+ * kernel's always-zero st_atime instead). Found via `ls -l` reporting
+ * every file as size 0 despite `cat`/the shell's own stat command (a
+ * direct in-kernel call, unaffected by this cross-boundary layout bug)
+ * showing the correct size. */
+typedef unsigned int nlink_t;
 
 struct stat {
     dev_t st_dev;
